tf to pt

+++ tf_utils.py
@@ -81,7 +81,7 @@ def noisy_label_loss_GCM(pred, cms, labels, alpha = 0.1):
         # loss_current = dice_loss2(pred_noisy, label_noisy.view(b, h, w).long())
         criterion = torch.nn.BCEWithLogitsLoss(reduce = 'mean')  # The loss function
         # Calculate the Loss
-        loss_current = criterion(pred_noisy, label_noisy.view(b, h, w).long())
+        loss_current = criterion(pred_noisy, label_noisy.view(b, h, w))
         main_loss += loss_current
         regularisation += torch.trace(torch.transpose(torch.sum(cm, dim = 0), 0, 1)).sum() / (b * h * w)
     #print("=====================")
